refactor: rename UsersModule to UserModule and update related imports
