+++ src/modules/sales/entities/sale.entity.ts
@@ -7,7 +7,7 @@ import {
   JoinColumn,
   OneToMany,
 } from 'typeorm';
-import { User } from '../../users/entities/user.entity';
+import { User } from '../../user/entities/user.entity';
 import { SaleItem } from './sale-item.entity';
 
 @Entity()
